requirements.txt
click
coloredlogs
cryptography
paramiko
fabric
PyYAML
invoke

pytest
SCAutolib
setuptools
schema
